cvo: refactor worker loops into their own files
